refactor(config,cli): deep restructure of @laufen/config + Result swap
Acts on a full audit of @laufen/config against the project's core
functional standards. Every kitchen-sink file is split into single-
purpose modules, side-effects are pushed to edges, every Result is now
the massaman object form `{ ok, value, error }`, and dead code is gone.

## @laufen/config restructure

```
src/
├── factory/{lauf,define-config,index}.ts   one type per file
├── config/
│   ├── types.ts                             LaufConfig, ResolvedLaufConfig, LoadedConfig
│   ├── schema.ts                            Zod schemas + DEFAULTS + validateConfig
│   ├── load.ts                              loadLaufConfig*, dedup'd via one canonical loader
│   ├── safe.ts                              two-line wrappers via attemptAsync
│   └── index.ts                             barrel
├── env/{dotenv,infisical,index}.ts          dotenv + infisical, fixed throw/console.warn
├── result.ts                                pure re-export of massaman/control
├── utils/json.ts                            uses massaman attempt, returns object Result
└── workspace/
    ├── config-files.ts                      shared CONFIG_FILE_NAMES + isRootConfig
    ├── package-name.ts                      readWorkspaceName
    ├── discover.ts                          discoverWorkspaces + findNearestWorkspace
    ├── root.ts                              resolveRoot — now consumes config-files
    └── scripts/
        ├── qualify.ts                       ROOT_WORKSPACE_NAME, qualify/extractStem
        ├── patterns.ts                      isValidPattern, stripScriptSuffix
        ├── canonical.ts                     canonicalize, filterToRoot
        ├── dedupe.ts                        dedupeByDeepestOwner
        ├── discover.ts                      discoverWorkspace/AllScripts
        ├── find.ts                          findScript
        └── index.ts                         barrel
```

Largest file shrinks: `config.ts` 268 LOC → split into 5 files averaging
~40 LOC. `workspace/scripts.ts` 264 LOC → split into 6 files averaging
~35 LOC. `workspace/discovery.ts` 171 LOC → split into 3 files.

## Fixed standard violations

- **`Promise.reject` → `throw`** in `env/dotenv.ts` and `env/infisical.ts`
  (unicorn/no-useless-promise-resolve-reject). The engine's
  `attemptAsync` still catches them at the boundary, but the rule
  prefers the direct form.
- **Removed `console.warn`** in dotenv: errors that aren't ENOENT now
  throw so callers can see them with full context (was silently warning
  and returning empty — actively misleading).
- **Wrapped `dotenv.parse()` in `attempt`**: could throw on malformed
  input.
- **Wrapped `execFileAsync` in `attempt`** in infisical: was throwing
  across the boundary unwrapped.
- **`Object.assign({}, ...results)` → `Object.fromEntries(flatMap)`**:
  no more mutating-style merge.
- **Raw `[new Error(...), null]` literals → `err(...)`/`ok(...)`** via
  the massaman helpers.
- **Deduplicated** `CONFIG_FILE_NAMES` and the `/root\s*:\s*true/` regex
  that lived in both `workspace/root.ts` and `workspace/discovery.ts`.
- **Collapsed** `loadLaufConfig` and `loadLaufConfigWithMeta` —
  `loadLaufConfig` is now a one-line wrapper. Same for
  `safeLoadLaufConfig` / `safeLoadLaufConfigWithMeta` — both are
  one-line `attemptAsync` wrappers.
- **Deleted dead code**: `src/types.ts` (4-line backwards-compat stub),
  `result.test.ts` (testing the local Result we no longer own).
- **Stray JSDoc** above `resolveFiles` removed in the dotenv rewrite.

## Result shape: tuple → massaman object

The whole `@laufen/config` surface (Result, ok, err, isOk, isErr,
attempt, attemptAsync, unwrap, safeParseJSON, safeLoad*) now uses
`{ ok, value, error }` discriminated objects from `massaman/control`.

The CLI consumers were updated:
- `assertOk` helper removed entirely
- `if (isErr(result)) { ctx.fail(...); return }` everywhere
- `result.value` for the happy path

Engine-side internals (resolveEnvValue, extractPackages, etc.) still
use the tuple form and are intentionally out of scope for this PR.
Two shapes coexist briefly at the CLI/engine boundary; that's the
follow-up.

## Barrels cleaned

- `index.ts` consolidates paired `export type {...}` + `export {...}`
- `workspace.ts` reduced to `export * from './workspace/index.ts'`
- `workspace/index.ts` exports the new file names

## Validation

- lint / fmt:check / typecheck / build — all green
- 448 tests pass (engine 299 + config 100 + cli 49)
- `lauf list` smoke verified through the bin shim
