+++ packages/cli/src/commands/blueprint.ts
@@ -3,8 +3,7 @@ import * as fs from 'node:fs';
 
 import type { CommandContext } from '@kidd-cli/core';
 import { command } from '@kidd-cli/core';
-import { assertOk, safeLoadLaufConfigWithMeta } from '@laufen/config';
-import { attempt } from 'es-toolkit';
+import { attempt, isErr, safeLoadLaufConfigWithMeta } from '@laufen/config';
 import * as path from 'pathe';
 import { z } from 'zod';
 
@@ -42,8 +41,11 @@ export default command({
     }
 
     const configResult = await safeLoadLaufConfigWithMeta(process.cwd());
-    assertOk(configResult, ctx.fail, 'Failed to load lauf config');
-    const loaded = configResult[1];
+    if (isErr(configResult)) {
+      ctx.fail(`Failed to load lauf config: ${configResult.error.message}`);
+      return;
+    }
+    const loaded = configResult.value;
 
     const targetDir = resolveTargetDir(ctx.args.dir, loaded.config.scripts, loaded.configDir);
     const normalizedTarget = path.normalize(targetDir);
@@ -64,28 +66,36 @@ export default command({
       ctx.fail(`File path escapes target directory: ${resolvedFilePath}`);
     }
 
-    const [mkdirError] = safeMkdirSync(targetDir);
-    if (mkdirError) {
-      ctx.fail(`Failed to create directory ${targetDir}: ${mkdirError}`);
+    const mkdir = safeMkdirSync(targetDir);
+    if (isErr(mkdir)) {
+      ctx.fail(`Failed to create directory ${targetDir}: ${safeParseError(mkdir.error)}`);
+      return;
     }
 
     const templateResult = getBlueprintTemplate(blueprintName);
-    assertOk(templateResult, ctx.fail, `Failed to load blueprint template "${blueprintName}"`);
-    const template = templateResult[1];
+    if (isErr(templateResult)) {
+      ctx.fail(
+        `Failed to load blueprint template "${blueprintName}": ${safeParseError(templateResult.error)}`,
+      );
+      return;
+    }
+    const template = templateResult.value;
 
-    const [writeError] = attempt(() =>
+    const written = attempt(() =>
       fs.writeFileSync(filePath, template, { encoding: 'utf-8', flag: 'wx' }),
     );
-    if (writeError) {
-      const nodeError = writeError as NodeJS.ErrnoException;
+    if (isErr(written)) {
+      const nodeError = written.error as NodeJS.ErrnoException;
       if (nodeError.code === 'EEXIST') {
         ctx.fail(`File already exists: ${filePath}`);
+        return;
       }
-      ctx.fail(`Failed to write ${filePath}: ${safeParseError(writeError)}`);
+      ctx.fail(`Failed to write ${filePath}: ${safeParseError(written.error)}`);
+      return;
     }
 
-    const [, pkg] = readPackageJSON(loaded.configDir);
-    const packageName = (pkg && pkg.name) || path.basename(loaded.configDir);
+    const pkg = readPackageJSON(loaded.configDir);
+    const packageName = (pkg.ok && pkg.value.name) || path.basename(loaded.configDir);
 
     const relative = path.relative(loaded.configDir, filePath);
     ctx.log.success(`Created ${relative}`);
