+++ packages/config/src/factory/index.ts
@@ -1,0 +1,2 @@
+export { defineConfig } from './define-config.ts';
+export { lauf } from './lauf.ts';
