--- packages/config/src/types.ts
@@ -1,4 +1,0 @@
-/**
- * Re-export DiscoveredScript from workspace module for backwards compatibility.
- */
-export type { DiscoveredScript } from './workspace/types.ts';
