+++ packages/config/src/workspace.ts
@@ -1,28 +1,7 @@
 /**
- * Workspace discovery surface used by the CLI.
+ * Workspace discovery surface used by `@laufen/cli`.
  *
- * Not part of the user-facing API — exposed as a sub-export for downstream
- * tooling (laufen CLI commands, tests) that needs to walk the workspace.
+ * Not part of the user-facing API — exposed as a sub-export for tooling
+ * that needs to walk the workspace.
  */
-export type {
-  CachedWorkspaceState,
-  DiscoveredScript,
-  Workspace,
-  WorkspaceRoot,
-  WorkspaceTree,
-} from './workspace/index.ts';
-export {
-  ROOT_WORKSPACE_NAME,
-  dedupeByDeepestOwner,
-  discoverAllScripts,
-  discoverWorkspaces,
-  discoverWorkspaceScripts,
-  findNearestWorkspace,
-  findScript,
-  getWorkspaceState,
-  qualifyScriptName,
-  resetWorkspaceCache,
-  resolveCurrentWorkspace,
-  resolveRoot,
-} from './workspace/index.ts';
-export { collectAncestors, walkUp } from './workspace/index.ts';
+export * from './workspace/index.ts';
