+++ packages/config/src/workspace/index.ts
@@ -1,6 +1,8 @@
 export type { CachedWorkspaceState } from './cache.ts';
 export { getWorkspaceState, resetWorkspaceCache } from './cache.ts';
-export { discoverWorkspaces, findNearestWorkspace } from './discovery.ts';
+export { CONFIG_FILE_NAMES, configNameFromFile, isRootConfig } from './config-files.ts';
+export { discoverWorkspaces, findNearestWorkspace } from './discover.ts';
+export { readWorkspaceName } from './package-name.ts';
 export { resolveCurrentWorkspace } from './resolve.ts';
 export { resolveRoot } from './root.ts';
 export {
@@ -10,6 +12,6 @@ export {
   discoverWorkspaceScripts,
   findScript,
   qualifyScriptName,
-} from './scripts.ts';
+} from './scripts/index.ts';
 export type { DiscoveredScript, Workspace, WorkspaceRoot, WorkspaceTree } from './types.ts';
 export { collectAncestors, walkUp } from './walk.ts';
