+++ packages/config/src/workspace/scripts/index.ts
@@ -1,0 +1,7 @@
+export { canonicalize, filterToRoot } from './canonical.ts';
+export { dedupeByDeepestOwner } from './dedupe.ts';
+export type { DiscoverAllOptions } from './discover.ts';
+export { discoverAllScripts, discoverWorkspaceScripts } from './discover.ts';
+export { findScript } from './find.ts';
+export { isValidPattern, stripScriptSuffix } from './patterns.ts';
+export { extractStem, qualifyScriptName, ROOT_WORKSPACE_NAME } from './qualify.ts';
